pyproject.toml
[project]
name = "my-mailer"
version = "0.1.0"
description = "A simple Flask API for handling contact form submissions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "flasgger>=0.9.7",
]
